+++ packages/formatters/index.js
@@ -29,8 +29,8 @@
  */
 
 module.exports = {
-  date: require('./date'),
-  duration: require('./duration'),
-  number: require('./number'),
-  time: require('./time')
+  date: require('./lib/date'),
+  duration: require('./lib/duration'),
+  number: require('./lib/number'),
+  time: require('./lib/time')
 };
