+++ packages/formatters/lib/duration.js
