+++ packages/formatters/lib/number.js
