+++ packages/formatters/lib/time.js
