+++ packages/formatters/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "messageformat-formatters",
+  "version": "2.0.0-rc.1",
+  "description": "Formatters for messageformat",
+  "keywords": [
+    "icu",
+    "messageformat",
+    "formatter"
+  ],
+  "author": "Eemeli Aro <eemeli@gmail.com>",
+  "license": "MIT",
+  "homepage": "https://messageformat.github.io/",
+  "repository": {
+    "type" : "git",
+    "url" : "https://github.com/messageformat/messageformat.git",
+    "directory": "packages/formatters"
+  },
+  "main": "index.js",
+  "files": [
+    "index.js",
+    "lib/"
+  ]
+}
