+++ packages/messageformat/package.json
@@ -49,6 +49,7 @@
   },
   "dependencies": {
     "make-plural": "^4.3.0",
+    "messageformat-formatters": "^2.0.0-rc.1",
     "messageformat-parser": "^4.1.0"
   },
   "devDependencies": {
